Styling and other fixes
1. Updated the styling across all html pages
2. Made some minor adjustments to the Listing model (max_length of description from 1000 to 256, winning_bidder allow blank=True)
3. Added img_url field for Category model
4. Changes to views:
a. In views.listing, we first check if a user is logged in before checking if a listing is in their watchlist
b. In create_listing, we no longer create a blank form and load into the page, since we manually created the form in HTML. Now we just load the list of categories.
c. In close_listing, we first check if there are any existing bids before we grab the 'winning_bidder'
d. In submit_bid, we first validate for invalid bid amounts being submitted